feat: 클래스 추출
1. 메서드를 분리하고 나면 update, insert 메서드에 중복된 코드가 보인다.
2. 각각 새로운 클래스로 분리해보자.
    1. UserDao에서 update 메서드와 관련된 코드를 UpdateJdbcTemplate 클래스를 만들어 이동시킨다.
    2. insert도 마찬가지로 InsertJdbcTemplate 클래스를 만들어 이동한다.
    3. UserDao에서 udpate, insert 각 메서드는 코드를 이동한 클래스를 사용하도록 수정한다.
    4. UpdateJdbcTemplate, InsertJdbcTemplate로 옮긴 메서드는 User, UserDao 객체를 의존하고 있다.
