+++ app/src/main/java/com/techcourse/dao/InsertJdbcTemplate.java
@@ -1,0 +1,54 @@
+package com.techcourse.dao;
+
+import com.techcourse.domain.User;
+import java.sql.Connection;
+import java.sql.PreparedStatement;
+import java.sql.SQLException;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+public class InsertJdbcTemplate {
+
+    private static final Logger log = LoggerFactory.getLogger(InsertJdbcTemplate.class);
+
+    public void insert(final User user, final UserDao userDao) {
+        final String sql = createQueryForInsert();
+
+        Connection conn = null;
+        PreparedStatement pstmt = null;
+        try {
+            conn = userDao.getDataSource().getConnection();
+            pstmt = conn.prepareStatement(sql);
+            log.debug("query : {}", sql);
+            setValuesForInsert(user, pstmt);
+            pstmt.executeUpdate();
+        } catch (SQLException e) {
+            log.error(e.getMessage(), e);
+            throw new RuntimeException(e);
+        } finally {
+            try {
+                if (pstmt != null) {
+                    pstmt.close();
+                }
+            } catch (SQLException ignored) {
+            }
+
+            try {
+                if (conn != null) {
+                    conn.close();
+                }
+            } catch (SQLException ignored) {
+            }
+        }
+    }
+
+    private String createQueryForInsert() {
+        return "insert into users (account, password, email) values (?, ?, ?)";
+    }
+
+    private void setValuesForInsert(final User user, final PreparedStatement pstmt) throws SQLException {
+        pstmt.setString(1, user.getAccount());
+        pstmt.setString(2, user.getPassword());
+        pstmt.setString(3, user.getEmail());
+    }
+}
